OK-376: Siirretty OutlinedCheckboxButton omaan tiedostoonsa
